refactor: constimpl package
