+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/collection/ext/CollectionExt.kt
@@ -1,10 +1,10 @@
 package xyz.lbres.kotlinutils.collection.ext
 
 import xyz.lbres.kotlinutils.iterable.ext.countElement
-import xyz.lbres.kotlinutils.set.multiset.ConstMultiSetImpl
-import xyz.lbres.kotlinutils.set.multiset.ConstMutableMultiSetImpl
 import xyz.lbres.kotlinutils.set.multiset.MultiSet
 import xyz.lbres.kotlinutils.set.multiset.MutableMultiSet
+import xyz.lbres.kotlinutils.set.multiset.constimpl.ConstMultiSetImpl
+import xyz.lbres.kotlinutils.set.multiset.constimpl.ConstMutableMultiSetImpl
 import xyz.lbres.kotlinutils.set.multiset.impl.MultiSetImpl
 import xyz.lbres.kotlinutils.set.multiset.impl.MutableMultiSetImpl
