+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/constimpl/ConstMultiSetImpl.kt
@@ -1,5 +1,6 @@
-package xyz.lbres.kotlinutils.set.multiset
+package xyz.lbres.kotlinutils.set.multiset.constimpl
 
+import xyz.lbres.kotlinutils.set.multiset.ConstMultiSet
 import xyz.lbres.kotlinutils.set.multiset.manager.ConstMultiSetManager
 
 /**
