+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/constimpl/ConstMutableMultiSetImplTest.kt
@@ -1,9 +1,9 @@
-package xyz.lbres.kotlinutils.set.multiset.impl.constimpl
+package xyz.lbres.kotlinutils.set.multiset.constimpl
 
 import xyz.lbres.kotlinutils.list.IntList
 import xyz.lbres.kotlinutils.set.multiset.ConstMutableMultiSet
 import xyz.lbres.kotlinutils.set.multiset.constMutableMultiSetOf
-import xyz.lbres.kotlinutils.set.multiset.impl.constimpl.mutable.* // ktlint-disable no-wildcard-imports no-unused-imports
+import xyz.lbres.kotlinutils.set.multiset.constimpl.mutable.* // ktlint-disable no-wildcard-imports no-unused-imports
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
