+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/constimpl/immutable/runConstUnaryTests.kt
@@ -1,4 +1,4 @@
-package xyz.lbres.kotlinutils.set.multiset.impl.constimpl.mutable
+package xyz.lbres.kotlinutils.set.multiset.constimpl.immutable
 
 import xyz.lbres.kotlinutils.list.IntList
 import xyz.lbres.kotlinutils.set.multiset.* // ktlint-disable no-wildcard-imports no-unused-imports
@@ -6,54 +6,33 @@ import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertTrue
 
-fun runMutableConstIsEmptyTests() {
+fun runConstIsEmptyTests() {
     // empty
-    var intSet: ConstMutableMultiSet<Int> = constMutableMultiSetOf()
+    var intSet: ConstMultiSet<Int> = constMultiSetOf()
     assertTrue(intSet.isEmpty())
 
-    var stringSet: ConstMutableMultiSet<String> = constMutableMultiSetOf()
+    var stringSet: ConstMultiSet<String> = constMultiSetOf()
     assertTrue(stringSet.isEmpty())
 
     // not empty
-    intSet = constMutableMultiSetOf(0)
+    intSet = constMultiSetOf(0)
     assertFalse(intSet.isEmpty())
 
-    intSet = constMutableMultiSetOf(1000, -1000, 4, 2, 4)
+    intSet = constMultiSetOf(1000, -1000, 4, 2, 4)
     assertFalse(intSet.isEmpty())
 
-    intSet = constMutableMultiSetOf(3, 3, 3)
+    intSet = constMultiSetOf(3, 3, 3)
     assertFalse(intSet.isEmpty())
 
-    stringSet = constMutableMultiSetOf("123", "abc")
+    stringSet = constMultiSetOf("123", "abc")
     assertFalse(stringSet.isEmpty())
 
-    stringSet = constMutableMultiSetOf("hello world", "hello world")
+    stringSet = constMultiSetOf("abcdefg", "abcdefg")
     assertFalse(stringSet.isEmpty())
-
-    // remove elements
-    intSet = constMutableMultiSetOf(1)
-    intSet.remove(1)
-    assertTrue(intSet.isEmpty())
-
-    intSet = constMutableMultiSetOf(1, 1)
-    intSet.remove(1)
-    assertFalse(intSet.isEmpty())
-    intSet.remove(1)
-    assertTrue(intSet.isEmpty())
-
-    intSet = constMutableMultiSetOf(2, 3)
-    intSet.remove(3)
-    assertFalse(intSet.isEmpty())
-    intSet.remove(2)
-    assertTrue(intSet.isEmpty())
-
-    intSet = constMutableMultiSetOf(2, 3)
-    intSet.clear()
-    assertTrue(intSet.isEmpty())
 }
 
-fun runMutableConstGetCountOfTests() {
-    var set: ConstMutableMultiSet<Int> = constMutableMultiSetOf()
+fun runConstGetCountOfTests() {
+    var set: ConstMultiSet<Int> = constMultiSetOf()
     var expected = 0
 
     var value = 0
@@ -62,7 +41,7 @@ fun runMutableConstGetCountOfTests() {
     value = 100
     assertEquals(expected, set.getCountOf(value))
 
-    set = constMutableMultiSetOf(2)
+    set = constMultiSetOf(2)
 
     value = 2
     expected = 1
@@ -72,7 +51,7 @@ fun runMutableConstGetCountOfTests() {
     expected = 0
     assertEquals(expected, set.getCountOf(value))
 
-    set = constMutableMultiSetOf(1, 1, 2, 1, -4, 5, 2)
+    set = constMultiSetOf(1, 1, 2, 1, -4, 5, 2)
 
     value = 1
     expected = 3
@@ -90,9 +69,9 @@ fun runMutableConstGetCountOfTests() {
     expected = 1
     assertEquals(expected, set.getCountOf(value))
 
-    val list1 = listOf(1, 2, 3)
-    val list2 = listOf(1, 2, 3)
-    val listSet: ConstMutableMultiSet<IntList> = constMutableMultiSetOf(list1, list2)
+    val list1 = mutableListOf(1, 2, 3)
+    val list2 = mutableListOf(1, 2, 3)
+    val listSet: ConstMultiSet<IntList> = constMultiSetOf(list1, list2)
 
     var listValue = listOf(1, 2, 3)
     expected = 2
